+++ js-jest/src/item_updaters/aged_brie_updater.js
@@ -1,4 +1,4 @@
-const { ItemUpdater } = require("../gilded_rose");
+const { ItemUpdater } = require("../item_updater");
 
 class AgedBrieUpdater extends ItemUpdater {
     updateQuality() {
