+++ js-jest/src/item_updaters/backstage_passes_updater.js
@@ -1,4 +1,4 @@
-const { ItemUpdater } = require("../gilded_rose");
+const { ItemUpdater } = require("../item_updater");
 
 class BackStagePassesUpdater extends ItemUpdater {
     updateQuality() {
